feat(register): Added register feature

+++ src/runtime/composables/useFortifyFeatures.ts
@@ -23,6 +23,7 @@ export interface FortifyFeatures {
     twoFactorCredentials: TwoFactorCredentials
   ) => Promise<void>;
   disableTwoFactorAuthentication: () => Promise<void>;
+  register: (registrationCredentials: RegistrationCredentials) => Promise<void>;
 }
 
 // Define the login credentials type
@@ -39,6 +40,13 @@ interface HasRecoveryCode {
   recovery_code: string;
 }
 type TwoFactorCredentials = HasCode | HasRecoveryCode;
+
+// Define the registeration credentials type
+interface RegistrationCredentials {
+  name: string;
+  password: string;
+  password_confirmation: string;
+}
 
 /**
  * This function initializes and returns the fortify features.
@@ -153,27 +161,53 @@ export function useFortifyFeatures(): FortifyFeatures {
    * Enable 2FA for the user
    *
    * Sends a POST request to enable 2FA for the user
+   * @see https://laravel.com/docs/11.x/fortify#enabling-two-factor-authentication
    *
    * @return {Promise<void>}
    */
   const enableTwoFactorAuthentication = async (): Promise<void> => {
+    // Check if the 2FA feature is enabled in the config
+    if (!config.features?.twoFactorAuthentication) {
+      throw new Error("2FA feature not enabled. Please enable it from config");
+    }
+
+    // Check if the 2fa enable endpoint is set in the config
+    if (!config.endpoints.tfa?.enable) {
+      throw new Error(
+        "2FA enable endpoint not set. Please set this endpoint from config"
+      );
+    }
+
     try {
-      return await api(config.endpoints.tfa?.enable!, { method: "POST" });
+      return await api(config.endpoints.tfa?.enable, { method: "POST" });
     } catch (error) {
       console.log(error);
     }
   };
 
   /**
-   * Fetches the QR code for two factor authentication.
+   * Fetches the QR code for two factor authentication setup.
    *
    * Sends a GET request to the server to fetch the QR code for two factor authentication.
+   * @see https://laravel.com/docs/11.x/fortify#enabling-two-factor-authentication
    *
    * @return {Promise<string>} The QR code svg data as a string.
    */
   const getTwoFactorAuthenticationQRCode = async (): Promise<void> => {
+    // Check if the 2FA feature is enabled in the config
+    if (!config.features?.twoFactorAuthentication) {
+      throw new Error("2FA feature not enabled. Please enable it from config");
+    }
+
+    // Check if the 2fa setup QR code endpoint is set in the config
+    if (!config.endpoints.tfa?.code) {
+      throw new Error(
+        "2FA code endpoint not set. Please set this endpoint from config"
+      );
+    }
+
     try {
-      return await api(config.endpoints.tfa?.code!, { method: "GET" });
+      return await api(config.endpoints.tfa?.code, { method: "GET" });
     } catch (error) {
       console.log(error);
     }
@@ -183,12 +217,25 @@ export function useFortifyFeatures(): FortifyFeatures {
    * Shows the two factor authentication recovery codes.
    *
    * Fetches the two factor authentication recovery codes from the server.
+   * @see https://laravel.com/docs/11.x/fortify#displaying-the-recovery-codes
    *
    * @return {Promise<void>} - A Promise that resolves when the recovery codes are fetched.
    */
   const showTwoFactorAuthenticationRecoveryCodes = async (): Promise<void> => {
+    // Check if the 2FA feature is enabled in the config
+    if (!config.features?.twoFactorAuthentication) {
+      throw new Error("2FA feature not enabled. Please enable it from config");
+    }
+
+    // Check if the 2fa recodery code endpoint is set in the config
+    if (!config.endpoints.tfa?.recoveryCode) {
+      throw new Error(
+        "2FA recovery code endpoint not set. Please set this endpoint from config"
+      );
+    }
+
     try {
-      return await api(config.endpoints.tfa?.recoveryCode!, { method: "GET" });
+      return await api(config.endpoints.tfa?.recoveryCode, { method: "GET" });
     } catch (error) {
       console.log(error);
     }
@@ -198,6 +245,7 @@ export function useFortifyFeatures(): FortifyFeatures {
    * Solves the two factor authentication challenge.
    *
    * Sends a POST request to the server to solve the two factor authentication challenge.
+   * @see https://laravel.com/docs/11.x/fortify#authenticating-with-two-factor-authentication
    *
    * @param {TwoFactorCredentials} twoFactorCredentials - The two factor authentication credentials.
    * @return {Promise<void>} - A Promise that resolves when the challenge is solved.
@@ -205,8 +253,20 @@ export function useFortifyFeatures(): FortifyFeatures {
   const solveTwoFactorAuthenticationChallenge = async (
     twoFactorCredentials: TwoFactorCredentials
   ): Promise<void> => {
+    // Check if the 2FA feature is enabled in the config
+    if (!config.features?.twoFactorAuthentication) {
+      throw new Error("2FA feature not enabled. Please enable it from config");
+    }
+
+    // Check if the 2fa challenge endpoint is set in the config
+    if (!config.endpoints.tfa?.challenge) {
+      throw new Error(
+        "2FA challenge endpoint not set. Please set this endpoint from config"
+      );
+    }
+
     try {
-      return await api(config.endpoints.tfa?.challenge!, {
+      return await api(config.endpoints.tfa?.challenge, {
         method: "POST",
         body: twoFactorCredentials,
       });
@@ -219,12 +279,64 @@ export function useFortifyFeatures(): FortifyFeatures {
    * Disables two-factor authentication.
    *
    * Sends a DELETE request to the server to disable two-factor authentication.
+   * @see https://laravel.com/docs/11.x/fortify#disabling-two-factor-authentication
    *
    * @return {Promise<void>} - A Promise that resolves when two-factor authentication is disabled.
    */
   const disableTwoFactorAuthentication = async (): Promise<void> => {
+    // Check if the 2FA feature is enabled in the config
+    if (!config.features?.twoFactorAuthentication) {
+      throw new Error("2FA feature not enabled. Please enable it from config");
+    }
+
+    // Check if the 2fa disable endpoint is set in the config
+    if (!config.endpoints.tfa?.disable) {
+      throw new Error(
+        "2FA disable endpoint not set. Please set this endpoint from config"
+      );
+    }
+
     try {
       await api(config.endpoints.tfa?.disable!, { method: "DELETE" });
+    } catch (error) {
+      console.log(error);
+    }
+  };
+
+  /**
+   * Registers a new user.
+   *
+   * Sends a POST request to the server to register a new user.
+   *
+   * @param {RegistrationCredentials} registrationCredentials - The registration credentials.
+   * @throws {Error} If the registration feature is not enabled in the config, or
+   * if the registration endpoint is not set in the config.
+   * @see https://laravel.com/docs/11.x/fortify#registration
+   * @return {Promise<void>} - A Promise that resolves when the user is registered.
+   */
+  const register = async (
+    registrationCredentials: RegistrationCredentials
+  ): Promise<void> => {
+    // Check if the registration feature is enabled in the config
+    if (!config.features?.registration) {
+      throw new Error(
+        "Registration feature not enabled. Please enable it from config"
+      );
+    }
+
+    // Check if the register endpoint is set in the config
+    if (!config.endpoints?.register) {
+      throw new Error(
+        "Register endpoint not set. Please set this endpoint from config"
+      );
+    }
+
+    try {
+      // Send a POST request to the server to register a new user
+      return await api(config.endpoints.register as RequestInfo, {
+        method: "POST",
+        body: registrationCredentials,
+      });
     } catch (error) {
       console.log(error);
     }
@@ -238,5 +350,6 @@ export function useFortifyFeatures(): FortifyFeatures {
     showTwoFactorAuthenticationRecoveryCodes,
     solveTwoFactorAuthenticationChallenge,
     disableTwoFactorAuthentication,
+    register,
   };
 }
